updated home card
