+++ screens/Home/homeStyle.tsx
@@ -1,0 +1,649 @@
+import { StyleSheet } from "react-native";
+
+const homeStyles = StyleSheet.create({
+  scrollview: {
+    alignItems: "center",
+    justifyContent: "center",
+    backgroundColor: "#000",
+  },
+
+  HomeCardContainer: {
+    alignItems: "center",
+    justifyContent: "center",
+    position: "relative",
+    width: "100%",
+  },
+
+  HomeCardNavBarContainer: {
+    alignItems: "center",
+    justifyContent: "center",
+    position: "relative",
+    width: "100%",
+    backgroundColor: "#000000",
+  },
+
+  HomeCardNavBarContent: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "space-between",
+    position: "relative",
+    color: "#FFFFFF",
+    width: 339.11,
+    paddingHorizontal: 22,
+  },
+
+  HomeCardNavText: {
+    fontStyle: "normal",
+    fontSize: 16,
+    lineHeight: 26,
+    color: "#FFFFFF",
+  },
+
+  categoryContainer: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "flex-start",
+    color: "#FFFFFF",
+  },
+
+  CategoryText: {
+    fontStyle: "normal",
+    fontSize: 16,
+    lineHeight: 26,
+    color: "#FFFFFF",
+  },
+
+  caretDownIcon: {
+    marginLeft: 22,
+    width: 14.55,
+    resizeMode: "contain",
+  },
+
+  HomeCardContent: {
+    alignItems: "center",
+    justifyContent: "center",
+    backgroundColor: "red",
+    position: "relative",
+    width: "100%",
+    // height: 762.6290893554688,
+    height: 603.99,
+  },
+
+  mainPoster: {
+    alignItems: "center",
+    justifyContent: "center",
+    position: "relative",
+    width: "100%",
+    // height: 762.6290893554688,
+    height: 603.99,
+  },
+
+  mainPosterActions: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "space-between",
+    width: 339.11,
+    height: 54.38,
+    position: "absolute",
+    top: 529.4,
+  },
+
+  myListContainer: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 51,
+  },
+
+  plus: {
+    width: 29.11,
+    height: 29.11,
+  },
+
+  myListText: {
+    fontSize: 15,
+    lineHeight: 18,
+    letterSpacing: 0,
+    textAlign: "center",
+    color: "#FFFFFF",
+  },
+
+  mainPlayButton: {
+    width: 119.34,
+    height: 46.57,
+  },
+
+  infoContainer: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 29.11,
+    height: 54.38,
+  },
+
+  infoIcon: {
+    width: 29.11,
+    height: 29.11,
+  },
+
+  infoText: {
+    fontSize: 14.55,
+    lineHeight: 17.61,
+    textAlign: "center",
+    color: "#FFFFFF",
+  },
+
+  netflix_sans_loader: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    position: "relative",
+  },
+
+  netflix_dir_ltr: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    position: "relative",
+  },
+
+  netflix_dark_background: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    position: "relative",
+  },
+
+  faUserBell: {
+    color: "#fff",
+  },
+
+  mainView: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    position: "relative",
+    zIndex: 0,
+  },
+
+  lolomo_fullbleed: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    paddingBottom: 50,
+    zIndex: 0,
+  },
+
+  volatile_billboard: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    zIndex: 1,
+  },
+
+  billboard_row: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    height: 150,
+    marginBottom: 20,
+    paddingBottom: "40%",
+    backgroundColor: "#000",
+  },
+
+  billboard_billboardPane: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    position: "absolute",
+    top: 0,
+  },
+
+  hero_image_wrapper: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    height: 210.94,
+  },
+
+  info_meta_layer: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 277,
+    position: "absolute",
+    left: "4%",
+    top: 0,
+    zIndex: 10,
+  },
+
+  logo_text__meta_layer: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 277,
+  },
+
+  billboard_title: {
+    alignItems: "flex-start",
+    justifyContent: "center",
+    width: 277,
+    height: 60.14,
+    marginBottom: 4.5,
+  },
+
+  info_wrapper: {
+    alignItems: "flex-start",
+    justifyContent: "center",
+    width: 277,
+  },
+
+  synopsis_fade_container: {
+    alignItems: "flex-start",
+    justifyContent: "center",
+    width: 277,
+  },
+
+  synopsis_no_supplemental: {
+    alignItems: "flex-start",
+    justifyContent: "center",
+    width: 277,
+    marginTop: 9.6,
+  },
+
+  synopsis_track_text: {
+    fontSize: 9,
+    color: "#fff",
+  },
+
+  billboard_links_player: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "flex-start",
+    width: 277,
+    marginTop: 28.8,
+  },
+
+  playButton: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    width: 67.48,
+    height: 28.13,
+    borderTopRightRadius: 4,
+    borderTopLeftRadius: 4,
+    borderBottomLeftRadius: 4,
+    borderBottomRightRadius: 4,
+    marginBottom: 5.625,
+    marginRight: 5.625,
+    paddingTop: 4.5,
+    paddingRight: 13.5,
+    paddingBottom: 4.5,
+    paddingLeft: 11.25,
+    backgroundColor: "#fff",
+  },
+
+  faUserPlay: {
+    color: "#000",
+  },
+
+  playButtonText: {
+    color: "#000",
+    marginLeft: 4,
+    fontSize: 9,
+  },
+
+  moreButton: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    width: 84.82,
+    height: 28.13,
+    borderTopRightRadius: 4,
+    borderTopLeftRadius: 4,
+    borderBottomLeftRadius: 4,
+    borderBottomRightRadius: 4,
+    marginBottom: 5.625,
+    marginRight: 5.625,
+    paddingTop: 4.5,
+    paddingRight: 13.5,
+    paddingBottom: 4.5,
+    paddingLeft: 11.25,
+    backgroundColor: "#6d6d6eb3",
+  },
+
+  faUserInfo: {
+    color: "#fff",
+  },
+
+  moreButtonText: {
+    fontSize: 9,
+    color: "#fff",
+    marginLeft: 4,
+  },
+
+  embeded_components: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    height: 41,
+    position: "absolute",
+    right: 0,
+    bottom: 6,
+  },
+
+  actionButtons: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 31.13,
+    height: 31.13,
+    borderWidth: 1,
+    borderColor: "rgba(255, 255, 255, 0.7)",
+    borderRadius: 22,
+    padding: 4.5,
+    marginRight: 8,
+    position: "absolute",
+    bottom: 0,
+    right: 32,
+  },
+
+  maturity_rating: {
+    alignItems: "center",
+    justifyContent: "center",
+    paddingVertical: 1.87,
+    paddingRight: 13.125,
+    paddingLeft: 3,
+    borderLeftWidth: 3,
+    borderLeftColor: "#fff",
+    backgroundColor: "#33333399",
+    position: "absolute",
+    right: 0,
+    top: 2,
+  },
+
+  maturity_number: {
+    fontSize: 9,
+    color: "#fff",
+  },
+
+  lolomoRow_title_card: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    marginTop: 57.6,
+    marginBottom: 19.2,
+  },
+
+  rowHeader_ltr: {
+    alignItems: "flex-start",
+    justifyContent: "center",
+    width: 375,
+  },
+
+  rowTitle_ltr0: {
+    alignItems: "flex-start",
+    justifyContent: "center",
+    width: 191.05,
+    marginHorizontal: "4%",
+    marginBottom: 8,
+    fontWeight: 700,
+  },
+
+  row_header_title: {
+    fontSize: 12,
+    color: "#e5e5e5",
+  },
+
+  rowContainer_title_card: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+  },
+
+  ptrack_container: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    height: 123.66,
+  },
+
+  rowContent_slider: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    height: 123.66,
+  },
+
+  slider: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    height: 123.66,
+    paddingHorizontal: "4%",
+  },
+
+  handlePrev: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 15,
+    height: 123.66,
+    backgroundColor: "hsla(0,0%,8%,.5)",
+    borderBottomRightRadius: 4,
+    borderTopRightRadius: 4,
+    position: "absolute",
+    left: 10,
+  },
+
+  sliderMask_showPeek: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 345,
+  },
+
+  handleNext: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 15,
+    height: 123.66,
+    backgroundColor: "hsla(0,0%,8%,.5)",
+    borderBottomLeftRadius: 4,
+    borderTopLeftRadius: 4,
+    position: "absolute",
+    right: 10,
+  },
+
+  sliderContent: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    width: 345,
+    height: 122.66,
+  },
+
+  animated_container: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    width: 345,
+  },
+
+  sliderItem: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 172.5,
+    height: 122.66,
+    paddingHorizontal: 3.84,
+  },
+
+  titleCard_Container: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 171.75,
+  },
+
+  boxArt_size: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    width: 171.75,
+    paddingVertical: "35.714285714%",
+    borderRadius: 3.84,
+  },
+
+  rank_1_container: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 87.88,
+    height: 124.66,
+  },
+
+  lolomoRow_title_card_1: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    marginBottom: 19.2,
+  },
+
+  rowContainer_title_card_1: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    height: 97.59,
+  },
+
+  ptrack_container_1: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    height: 97.59,
+  },
+
+  rowContent_slider_1: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    height: 97.59,
+  },
+
+  slider_1: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    height: 97.59,
+    paddingHorizontal: "4%",
+  },
+
+  sliderItem1: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 172.5,
+    height: 96.59,
+  },
+
+  titleCard_Container1: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 171.75,
+    height: 96.59,
+  },
+
+  boxArt_size1: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    paddingVertical: "28.125%",
+    width: 171.75,
+    height: 96.59,
+    borderRadius: 3.84,
+  },
+
+  lolomoRow_title_card_2: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    marginBottom: 19.2,
+  },
+
+  contentInfo_footer: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 375,
+    marginTop: 20,
+    paddingHorizontal: "4%",
+  },
+
+  social_links: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "flex-start",
+    width: 345,
+    marginBottom: 16,
+  },
+
+  social_links_icon: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    width: 36,
+    height: 29.8,
+    marginRight: 15,
+  },
+
+  social_icon_container: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 38,
+    height: 27,
+  },
+
+  facebook: {
+    color: "white",
+  },
+
+  social_link_list: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 345,
+    marginBottom: 14,
+  },
+
+  member_footer_link_wrapper: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    width: 345,
+  },
+
+  member_footer_links: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "flex-start",
+    width: 172.5,
+    marginBottom: 16,
+  },
+
+  member_footer_link_text: {
+    color: "#737373",
+    fontSize: 13,
+  },
+
+  member_footer_service: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "flex-start",
+    width: 345,
+  },
+
+  member_footer_button: {
+    alignItems: "center",
+    justifyContent: "center",
+    width: 94.33,
+    padding: 8,
+    borderWidth: 1,
+    borderColor: "#737373",
+  },
+
+  member_footer_btn: {
+    fontSize: 13,
+    color: "#737373",
+  },
+});
+
+export default homeStyles;
